requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
newspaper3k==0.2.8
nltk==3.9.2
pillow==11.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
